requirements.txt
asgiref==3.11.1
async-timeout==5.0.1
Django==5.2.11
django-filter==25.2
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==25.1.0
packaging==26.0
PyJWT==2.11.0
redis==7.2.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
